Almost final commit for version 2.0
git-svn-id: svn+ssh://oberon/svn/OMiSCID-Dev@9302 c4bd1236-c4f7-0310-a2fe-a60683d9aa4a

+++ Json/Json/json_spirit_value.h
@@ -54,7 +54,7 @@ namespace json_spirit
         inline void get_val(bool& b) const {b = get_bool();}
         inline void get_val(int& i) const {i = get_int();}
         inline void get_val(double& d) const {d = get_real();}
-        inline void get_val(float& f) const {f = get_real();}
+        inline void get_val(float& f) const {f = (float)get_real();}
         
         Object& get_obj();
         Array&  get_array();
